rename core & fix tests
