--- eth_vertigo/mutation/__init__.py
@@ -1,1 +1,0 @@
-from eth_vertigo.mutation.mutation import Mutation, MutationResult
