--- eth_vertigo/mutation/templates/mutation_template.jinja2
@@ -1,8 +1,0 @@
-Mutation:
-    File: {{ file_name }}
-    Line nr: {{ line_number }}
-    Result: {{ result }}
-    Original line:
-         {{ original_line }}
-    Mutated line:
-         {{ mutated_line }}
